+++ docs/assets/formatter-EIJCOSYU-QmdoZi-G.js
@@ -1,0 +1,1 @@
+import{d as m,r as o}from"./DocsRenderer-LL677BLK-Bo2qiop2.js";import{b as t}from"./iframe-C8vF1WH5.js";import"./preload-helper-C1FmrZbK.js";import"./react-18-Cn0Hx7ie.js";import"./jsx-runtime-BjG_zV1W.js";var e=t(o(),1),_=(0,e.default)(2)(async(i,r)=>i===!1?r:m(r));export{_ as formatter};